content/base/test/xcsp/test_CSP_inlinestyle.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for Content Security Policy inline stylesheets stuff</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">
</div>

<iframe style="width:100%;height:300px;" id='cspframe'></iframe>
<script class="testbody" type="text/javascript">

//////////////////////////////////////////////////////////////////////
// set up and go
SimpleTest.waitForExplicitFinish();

var done = 0;

// Our original CSP implementation does not block inline styles.
function checkStyles(evt) {
  var cspframe = document.getElementById('cspframe');
  var color;

  // black means the style wasn't applied.  green colors are used for styles
  //expected to be applied.  A color is red if a style is erroneously applied
  color = window.getComputedStyle(cspframe.contentDocument.getElementById('linkstylediv'),null)['color'];
  ok('rgb(0, 255, 0)' === color, 'External Stylesheet (original CSP implementation) (' + color + ')');
  color = window.getComputedStyle(cspframe.contentDocument.getElementById('inlinestylediv'),null)['color'];
  ok('rgb(0, 255, 0)' === color, 'Inline Style TAG (original CSP implementation) (' + color + ')');
  color = window.getComputedStyle(cspframe.contentDocument.getElementById('attrstylediv'),null)['color'];
  ok('rgb(0, 255, 0)' === color, 'Style Attribute (original CSP implementation) (' + color + ')');
  // SMIL tests
  color = window.getComputedStyle(cspframe.contentDocument.getElementById('xmlTest',null))['fill'];
  ok('rgb(0, 255, 0)' === color, 'XML Attribute styling (SMIL) (' + color + ')');
  color = window.getComputedStyle(cspframe.contentDocument.getElementById('cssOverrideTest',null))['fill'];
  ok('rgb(0, 255, 0)' === color, 'CSS Override styling (SMIL) (' + color + ')');
  color = window.getComputedStyle(cspframe.contentDocument.getElementById('cssOverrideTestById',null))['fill'];
  ok('rgb(0, 255, 0)' === color, 'CSS Override styling via ID lookup (SMIL) (' + color + ')');
  color = window.getComputedStyle(cspframe.contentDocument.getElementById('cssSetTestById',null))['fill'];
  ok('rgb(0, 255, 0)' === color, 'CSS Set Element styling via ID lookup (SMIL) (' + color + ')');
  checkIfDone();
}

function checkIfDone() {
  done++;
  if (done == 1)
    SimpleTest.finish();
}

SpecialPowers.pushPrefEnv(
  {'set':[["security.csp.speccompliant", true]]},
  function() {
    // save this for last so that our listeners are registered.
    // ... this loads the testbed of good and bad requests.
    document.getElementById('cspframe').src = 'file_CSP_inlinestyle_main.html';
    document.getElementById('cspframe').addEventListener('load', checkStyles, false);
  }
);
</script>
</pre>
</body>
</html>
